+++ src/psionics.c
@@ -2921,7 +2921,7 @@ void spell_ether_warp(int level, P_char ch, char *arg, int type, P_char victim, 
     to_room = ch->in_room;
   }
 
-  distance = (int)(level * (int)get_property("spell.dim.perlevel.modifier", 1.35));
+  distance = (int)(level * get_property("spell.dim.perlevel.modifier", 1.35));
   if( !IS_TRUSTED(ch) && IS_NPC(victim) && (GET_SPEC(ch, CLASS_PSIONICIST, SPEC_PSYCHEPORTER) || IS_ILLITHID(ch))
     && how_close(ch->in_room, victim->in_room, distance) < 0
     && how_close(victim->in_room, ch->in_room, distance) < 0 )
@@ -3123,7 +3123,7 @@ void spell_wormhole(int level, P_char ch, char *arg, int type, P_char victim, P_
          success = false;
        else
        {
-         distance = (int)(level * (int)get_property("spell.dim.perlevel.modifier", 1.35));
+         distance = (int)(level * get_property("spell.dim.perlevel.modifier", 1.35));
 
          if( how_close(ch->in_room, to_room, distance) < 0
           && how_close(to_room, ch->in_room, distance) < 0 )
